add function module
