+++ tests/test_jacobians.py
@@ -1,6 +1,7 @@
 import pytest
 import torch
 
+from ncopt.functions.quadratic import Quadratic
 from ncopt.utils import (
     compute_batch_jacobian,
     compute_batch_jacobian_naive,
@@ -10,65 +11,6 @@ from ncopt.utils import (
 d = 10
 m = 3
 b = 4
-
-
-class Quadratic(torch.nn.Module):
-    def __init__(self, input_dim: int = None, params: tuple = None):
-        """Implements the function
-
-            x --> (1/2)*x.T*A*x + b.T*x + c
-
-        Parameters
-        ----------
-        input_dim : int, optional
-            If no params are specified this specifies the dimension of the tensors, by default None
-        params : tuple, optional
-           3-Tuple of tensors, by default None.
-           If specified, should contain the values of A, b and c.
-        """
-        super().__init__()
-
-        assert params is not None or (
-            input_dim is not None
-        ), "Specify either a dimension or parameters"
-        if params is not None:
-            assert (
-                len(params) == 3
-            ), f"params should contains three elements (A,b,c), but contains only {len(params)}."
-
-        if params is None:
-            self.A = torch.nn.Parameter(torch.randn(input_dim, input_dim))
-            self.b = torch.nn.Parameter(
-                torch.randn(
-                    input_dim,
-                )
-            )
-            self.c = torch.nn.Parameter(
-                torch.randn(
-                    1,
-                )
-            )
-        else:
-            self.A = torch.nn.Parameter(params[0])
-            self.b = torch.nn.Parameter(params[1])
-            self.c = torch.nn.Parameter(params[2])
-
-    def forward(self, x: torch.tensor) -> torch.tensor:
-        """
-
-        Parameters
-        ----------
-        x : torch.tensor
-            Should be batched, and have shape [batch_size, input_dim]
-
-        Returns
-        -------
-        torch.tensor of shape [batch_size, 1]
-        """
-        out = 0.5 * torch.sum((x @ self.A) * x, dim=1, keepdim=True)
-        out = out + (x * self.b).sum(dim=1, keepdim=True)
-        out = out + self.c
-        return out
 
 
 class DummyNet(torch.nn.Module):
@@ -89,19 +31,6 @@ class DummyNet(torch.nn.Module):
 
 ######################################################
 #### Tests start here
-
-
-def test_quadratic():
-    params = (torch.eye(d), torch.zeros(d), torch.tensor(1.0))
-    model = Quadratic(params=params)
-    inputs = torch.randn(b, d)
-
-    expected = 0.5 * (inputs * inputs).sum(1) + 1.0
-    out = model(inputs)
-
-    assert torch.allclose(out, expected[:, None])
-    assert out.shape == torch.Size([b, 1])
-    return
 
 
 @pytest.mark.parametrize(
